requirements.txt
Django==3.2.13
djangorestframework==3.13.1
djangorestframework-simplejwt==4.3.0
Pillow==9.5.0